+++ docs/Changelog.rst
@@ -3,8 +3,18 @@
 Changelog
 =========
 
+3.5.1 (21 March 2016)
+---------------------
+
+**Bug fixes**
+
+  * Add more logging information to snapshot delete method #582 (untergeek)
+  * Improve default timeout, logging, and exception handling for `seal` command
+    #583 (untergeek)
+  * Fix use of default snapshot name. #584 (untergeek)
+
 3.5.0 (16 March 2016)
-----------------------
+---------------------
 
 **General**